# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Support Chat API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  handle: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface UserProfileInput {
  /**
     * @minLength 3
     * @maxLength 30
     */
  handle: string;
  /** @minLength 1 */
  displayName: string;
  bio?: string;
}

export interface PublicProfile {
  handle: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface ChatLink {
  id: number;
  slug: string;
  label: string;
  createdAt: string;
}

export interface CreateLinkInput {
  /**
     * @minLength 3
     * @maxLength 50
     */
  slug: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  label: string;
}

export type ConversationStatus = typeof ConversationStatus[keyof typeof ConversationStatus];


export const ConversationStatus = {
  open: 'open',
  closed: 'closed',
} as const;

export interface Conversation {
  id: number;
  guestName: string;
  /** @nullable */
  guestEmail?: string | null;
  subject: string;
  status: ConversationStatus;
  isRead: boolean;
  lastMessageAt: string;
  messageCount: number;
  createdAt: string;
}

export type MessageSenderType = typeof MessageSenderType[keyof typeof MessageSenderType];


export const MessageSenderType = {
  guest: 'guest',
  owner: 'owner',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  senderType: MessageSenderType;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface ContactInput {
  /** @minLength 1 */
  guestName: string;
  /** @minLength 1 */
  message: string;
}

export interface GuestConversation {
  id: number;
  token: string;
  subject: string;
  createdAt: string;
}

export interface GuestConversationDetail {
  id: number;
  subject: string;
  guestName: string;
  ownerDisplayName: string;
  messages: Message[];
  createdAt: string;
}

export interface GuestMessageInput {
  /** @minLength 1 */
  content: string;
}

export interface ConversationStats {
  total: number;
  unread: number;
  open: number;
  closed: number;
}

export type DeleteLink200 = {
  success: boolean;
};