# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Support Chat API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: users
  - name: conversations
  - name: messages
  - name: links
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    put:
      operationId: upsertMe
      tags: [users]
      summary: Create or update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /users/{handle}:
    get:
      operationId: getUserByHandle
      tags: [users]
      summary: Get public user profile by handle
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: Not found

  /links:
    get:
      operationId: listLinks
      tags: [links]
      summary: List all custom chat links for authenticated user
      responses:
        "200":
          description: List of links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatLink"
    post:
      operationId: createLink
      tags: [links]
      summary: Create a new custom chat link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkInput"
      responses:
        "201":
          description: Created link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatLink"

  /links/{id}:
    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete a custom chat link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success

  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List all conversations for authenticated user
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"

  /conversations/{id}:
    get:
      operationId: getConversation
      tags: [conversations]
      summary: Get a conversation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
          description: Not found

  /conversations/{id}/read:
    post:
      operationId: markConversationRead
      tags: [conversations]
      summary: Mark a conversation as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a conversation (owner reply)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /public/{handle}/contact:
    post:
      operationId: startConversation
      tags: [conversations]
      summary: Start a conversation via handle (no auth)
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestConversation"

  /public/link/{slug}/contact:
    post:
      operationId: startConversationByLink
      tags: [conversations]
      summary: Start a conversation via custom link slug (no auth)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestConversation"
        "404":
          description: Link not found

  /public/link/{slug}:
    get:
      operationId: getLinkProfile
      tags: [links]
      summary: Get public profile for a custom link slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: Not found

  /public/conversations/{token}:
    get:
      operationId: getGuestConversation
      tags: [conversations]
      summary: Get conversation as guest via token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guest conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestConversationDetail"
        "404":
          description: Not found

  /public/conversations/{token}/messages:
    post:
      operationId: sendGuestMessage
      tags: [messages]
      summary: Guest sends a follow-up message
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /conversations/stats:
    get:
      operationId: getConversationStats
      tags: [conversations]
      summary: Get inbox stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        handle:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, clerkId, handle, displayName, createdAt]

    UserProfileInput:
      type: object
      properties:
        handle:
          type: string
          minLength: 3
          maxLength: 30
        displayName:
          type: string
          minLength: 1
        bio:
          type: string
      required: [handle, displayName]

    PublicProfile:
      type: object
      properties:
        handle:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
      required: [handle, displayName]

    ChatLink:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        label:
          type: string
        createdAt:
          type: string
      required: [id, slug, label, createdAt]

    CreateLinkInput:
      type: object
      properties:
        slug:
          type: string
          minLength: 3
          maxLength: 50
        label:
          type: string
          minLength: 1
          maxLength: 80
      required: [slug, label]

    Conversation:
      type: object
      properties:
        id:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: ["string", "null"]
        subject:
          type: string
        status:
          type: string
          enum: [open, closed]
        isRead:
          type: boolean
        lastMessageAt:
          type: string
        messageCount:
          type: integer
        createdAt:
          type: string
      required: [id, guestName, subject, status, isRead, lastMessageAt, messageCount, createdAt]

    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderType:
          type: string
          enum: [guest, owner]
        content:
          type: string
        createdAt:
          type: string
      required: [id, conversationId, senderType, content, createdAt]

    MessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required: [content]

    ContactInput:
      type: object
      properties:
        guestName:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
      required: [guestName, message]

    GuestConversation:
      type: object
      properties:
        id:
          type: integer
        token:
          type: string
        subject:
          type: string
        createdAt:
          type: string
      required: [id, token, subject, createdAt]

    GuestConversationDetail:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        guestName:
          type: string
        ownerDisplayName:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        createdAt:
          type: string
      required: [id, subject, guestName, ownerDisplayName, messages, createdAt]

    GuestMessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required: [content]

    ConversationStats:
      type: object
      properties:
        total:
          type: integer
        unread:
          type: integer
        open:
          type: integer
        closed:
          type: integer
      required: [total, unread, open, closed]