# lib/api-zod/src/generated/types/chatLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Support Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatLink {
  id: number;
  slug: string;
  label: string;
  createdAt: string;
}