# lib/api-zod/src/generated/types/createLinkInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Support Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLinkInput {
  /**
     * @minLength 3
     * @maxLength 50
     */
  slug: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  label: string;
}